+++ class-components/src/app/pages/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Home = () => {
+  return <div>Home page</div>;
+};
+
+export default Home;
