+++ class-components/src/app/pages/not-found.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NotFound = () => {
+  return <div>Page not founded</div>;
+};
+
+export default NotFound;
