+++ class-components/src/hooks/useLocalStorageQuery.ts
